// Semestre 5/DAWII/dawii_2022_1_grupo_02_domingo_frontend/src/app/models/producto.model.ts
import { Marca } from "./marca.model";
import { Pais } from "./pais.model";

export class Producto {
    
    idProducto?: number;
    nombre?: String;
    serie?: String;
    durabilidad?: String;
    fechaVigencia?: Date;
    precio?: number;
    stock?: number;
    marca?: Marca;
    pais?: Pais;
    estado?: number;
    fechaRegistro?: Date;
    mensaje?:any;

}